requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
asyncpg==0.29.0
pydantic==2.10.6
pydantic-settings==2.8.1
python-dotenv==1.0.1
aio-pika==9.5.5
alembic==1.15.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3        
openai==1.65.4
tenacity==9.0.0
Pillow==9.4.0
aiohttp==3.10.5
python-dateutil==2.8.2
minio==7.2.15
firebase-admin==6.0.0
httpx==0.24.1
redis==5.2.1
redis-om==0.3.3
sentry-sdk==1.39.1
PyJWT==2.10.1